requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Werkzeug==3.0.1
SQLAlchemy==2.0.23
WTForms==3.1.1
email-validator==2.1.0.post1
python-slugify==8.0.1
Pillow==10.1.0
Flask-Admin==1.6.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
gunicorn==21.2.0
